Refactored Local Groups to new format
